+++ NEWS.md
@@ -8,3 +8,4 @@
     * `randomForest::randomForest`
     * `lme4::lmer`
     * `quantreg::rq, nlrq, rqss, crq` (thanks to [Joran Elias](https://github.com/joranE))
+    * `lme4::glmer` (thanks to [Mathew Ling](https://github.com/Lingtax/))